project/requirements.txt
aioredis==1.2.0
amqp==2.5.0
asgiref==3.1.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.6.2
Automat==0.7.0
Babel==2.7.0
billiard==3.6.0.0
CacheControl==0.12.5
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
daphne==2.3.0
defusedxml==0.6.0
deprecation==2.0.7
Django==2.2.2
django-allauth==0.39.1
django-appconf==1.0.3
django-bootstrap3==11.0.0
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-celerybeat-status==0.0.10
django-cleanup==4.0.0
django-cors-headers==3.0.2
django-countries==5.3.3
django-crispy-forms==1.7.2
django-extensions==2.1.9
django-imagekit==4.0.2
django-paypal==1.0.0
django-redis==4.10.0
django-rest-swagger==2.2.0
django-stdimage==4.0.1
django-templated-mail==1.1.1
django-timezone-field==3.0
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djoser==1.7.0
entrypoints==0.3
flake8==3.7.7
flower==0.9.3
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10.1
jsonpickle==0.9.6
kombu==4.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.1
oauthlib==3.0.1
openapi-codec==1.3.2
packaging==19.2
pilkit==2.0
Pillow==6.0.0
progressbar2==3.42.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pyparsing==2.4.2
python-crontab==2.3.6
python-dateutil==2.8.0
python-dotenv==0.10.3
python-utils==2.3.0
python3-openid==3.1.0
pytz==2019.1
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
SQLAlchemy==1.3.4
sqlparse==0.3.0
squareup==3.20190925.0
tornado==5.1.1
Twisted==19.2.1
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.3
uWSGI==2.0.18
vine==1.3.0
zope.interface==4.6.0
